Homework_9/HW_10/requirement.txt
anyio==3.6.2
astroid==2.13.3
certifi==2022.12.7
colorama==0.4.6
dill==0.3.6
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==2.6.2
pylint==2.15.10
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
wrapt==1.14.1
